python/apps-version/src/requirements.txt
aiohttp
prometheus_async
pyyaml
prometheus_client
aiojobs
